backend/requirements.txt
# CPU优化版本 - 适合云端部署
Flask==2.3.3
Flask-CORS==4.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
Pillow==10.0.0
numpy==1.24.3
psutil==5.9.5
streamlit==1.28.0
matplotlib==3.7.2 